Major refactoring: move Executor trait method impls into trait itself, make 'exec' macro simpler
